refactor(parser): extract pb script engine into pbScript.ts
Move pb directive parsing, expression evaluation, script text parsing,
directive execution and request mutations out of parser.ts into
src/lib/pbScript.ts. parser.ts keeps temporary re-exports so existing
imports keep working.
